requirement.txt
langchain-core 
line-bot-sdk
fastapi
uvicorn
dotenv
langgraph
google-api-python-client
PyMuPDF 
openpyxl 
python-docx 
python-pptx
apscheduler
pymongo
langchain-community 
langchain-chroma 
langchain-huggingface 
chromadb 
cohere 
torch
mistralai
langchain-openai
langchain_mongodb